Yayyyyy level 1

+++ Assets/Scripts/KillBox.cs
@@ -2,6 +2,8 @@ using UnityEngine;
 
 public class KillBox : MonoBehaviour
 {
+    private void Start() { }
+
     private void OnTriggerEnter2D(Collider2D other)
     {
         KillPlayer(other.FindPlayer());
@@ -14,6 +16,7 @@ public class KillBox : MonoBehaviour
 
     private void KillPlayer(Player player)
     {
+        if(!enabled) return;
         if(!player) return;
         player.Kill();
     }
